// front-end/src/app/models/routeSegment.model.ts
// export interface IRouteSegment {
//     start: number;
//     distance: number;
//     condition: string;
// }

export class RouteSegment {
    start: number;
    distance: number;
    condition: string;
    constructor() {
        // start: number, distance: number, condition: string
        // this.start = start;
        // this.distance = distance;
        // this.condition = condition;
    }
}